Minor tidying - remove commented out code, pull out a little helper class
