+++ src/main/kotlin/io/libp2p/transport/ws/WebSocketServerInitializer.kt
@@ -1,8 +1,6 @@
 package io.libp2p.transport.ws
 
 import io.netty.channel.ChannelHandler
-import io.netty.channel.ChannelHandlerContext
-import io.netty.channel.ChannelInboundHandlerAdapter
 import io.netty.channel.ChannelInitializer
 import io.netty.channel.socket.SocketChannel
 import io.netty.handler.codec.http.websocketx.WebSocketServerProtocolHandler
@@ -21,16 +19,6 @@ internal class WebSocketServerInitializer(
         pipeline.addLast(HttpObjectAggregator(65536))
         pipeline.addLast(WebSocketServerCompressionHandler())
         pipeline.addLast(WebSocketServerProtocolHandler("/", null, true))
-        pipeline.addLast(object: ChannelInboundHandlerAdapter() {
-            override fun userEventTriggered(ctx: ChannelHandlerContext, evt: Any) {
-                if(evt is WebSocketServerProtocolHandler.HandshakeComplete) {
-                    ctx.pipeline().addLast(WebFrameCodec())
-                    ctx.pipeline().addLast(connectionBuilder)
-                    ctx.pipeline().remove(this)
-                    ctx.fireChannelActive()
-                }
-                super.userEventTriggered(ctx, evt)
-            }
-        })
+        pipeline.addLast(WebSocketServerHandshakeListener(connectionBuilder))
     } // initChannel
 } // WebSocketServerInitializer
