+++ src/main/kotlin/io/libp2p/transport/ws/WsTransport.kt
@@ -2,7 +2,11 @@ package io.libp2p.transport.ws
 
 import io.libp2p.core.InternalErrorException
 import io.libp2p.core.multiformats.Multiaddr
-import io.libp2p.core.multiformats.Protocol
+import io.libp2p.core.multiformats.Protocol.DNSADDR
+import io.libp2p.core.multiformats.Protocol.IP4
+import io.libp2p.core.multiformats.Protocol.IP6
+import io.libp2p.core.multiformats.Protocol.TCP
+import io.libp2p.core.multiformats.Protocol.WS
 import io.libp2p.transport.ConnectionUpgrader
 import io.libp2p.transport.implementation.ConnectionBuilder
 import io.libp2p.transport.implementation.NettyTransport
@@ -11,13 +15,18 @@ import java.net.Inet4Address
 import java.net.Inet6Address
 import java.net.InetSocketAddress
 
+/**
+ * The WS transport can establish libp2p connections
+ * via WebSockets endpoints.
+ */
 class WsTransport(
     upgrader: ConnectionUpgrader
 ) : NettyTransport(upgrader) {
+
     override fun handles(addr: Multiaddr): Boolean {
-        return (addr.has(Protocol.IP4) || addr.has(Protocol.IP6) || addr.has(Protocol.DNSADDR)) &&
-                addr.has(Protocol.TCP) &&
-                addr.has(Protocol.WS)
+        return (addr.has(IP4) || addr.has(IP6) || addr.has(DNSADDR)) &&
+                addr.has(TCP) &&
+                addr.has(WS)
     } // handles
 
     override fun serverTransportBuilder(
@@ -40,14 +49,14 @@ class WsTransport(
 
     override fun toMultiaddr(addr: InetSocketAddress): Multiaddr {
         val proto = when (addr.address) {
-            is Inet4Address -> Protocol.IP4
-            is Inet6Address -> Protocol.IP6
+            is Inet4Address -> IP4
+            is Inet6Address -> IP6
             else -> throw InternalErrorException("Unknown address type $addr")
         }
         return Multiaddr(listOf(
             proto to proto.addressToBytes(addr.address.hostAddress),
-            Protocol.TCP to Protocol.TCP.addressToBytes(addr.port.toString()),
-            Protocol.WS to ByteArray(0)
+            TCP to TCP.addressToBytes(addr.port.toString()),
+            WS to ByteArray(0)
         ))
     } // toMultiaddr
 } // class WsTransport
